+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/trixnity/messenger/internal/compose/view/scenes/twopane/TwoPaneSceneStrategyFactory.kt
@@ -4,6 +4,9 @@ import androidx.compose.material3.adaptive.currentWindowAdaptiveInfo
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
 import de.connect2x.trixnity.messenger.abi.TrixnityMessengerPrivateApi
+import de.connect2x.trixnity.messenger.compose.view.DI
+import de.connect2x.trixnity.messenger.compose.view.get
+import de.connect2x.trixnity.messenger.compose.view.getOrNull
 import de.connect2x.trixnity.messenger.internal.compose.view.scenes.SceneStrategyFactory
 
 @TrixnityMessengerPrivateApi
@@ -19,7 +22,7 @@ private class TwoPaneSceneStrategyFactoryImpl : TwoPaneSceneStrategyFactory {
     @Composable
     override fun <T : Any> create(): TwoPaneSceneStrategy<T> {
         val windowSizeClass = currentWindowAdaptiveInfo(true).windowSizeClass
-
-        return remember(windowSizeClass) { TwoPaneSceneStrategy(windowSizeClass) }
+        val placeholder = DI.getOrNull<TwoPaneScenePlaceholder>()
+        return remember(windowSizeClass, placeholder) { TwoPaneSceneStrategy(windowSizeClass, placeholder) }
     }
 }